favicon adicionado + refatorando pastas e arquivos pra maior organizacao
